% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpCompare}
\alias{ptpCompare}
\title{Compare two log files on a graph}
\usage{
ptpCompare(loga, logb, value, start, end, output)
}
\arguments{
\item{loga}{data frame read using ptplogRead}

\item{logb}{data frame read using ptplogRead}

\item{value}{column of the log file to compare
e.g. offset, delay, slave.to.master, master.to.slave}

\item{output}{an output file name for a PNG image}
}
\description{
Compare two log files on a graph
}

