% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpGraph}
\alias{ptpGraph}
\title{Graphs a dataframe returned by a call to ptplog}
\usage{
ptpGraph(logframe, value, start, end, output)
}
\arguments{
\item{logframe}{A data frame returned by ptpLogRead}

\item{value}{Graph only a specific value rather than the four major ones.}

\item{output}{An optional file name to place a PNG of the graph into}
}
\description{
Graphs a dataframe returned by a call to ptplog
}

