% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpHistogram}
\alias{ptpHistogram}
\title{Draw a simple histogram of some one variable in our log}
\usage{
ptpHistogram(log, start, end, output)
}
\description{
Draw a simple histogram of some one variable in our log
}
\examples{
ptpHistogram(foo$log$offset)

}

