% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpHistogramCompare}
\alias{ptpHistogramCompare}
\title{Compare two data sets to each other using a histogram}
\usage{
ptpHistogramCompare(loga, logb, starta, enda, startb, endb, output)
}
\description{
examples ptpHistogramCompare(foo$log$offset, bar$log$offset)
}

