\name{ptplib-package}
\alias{ptplib-package}
\alias{ptplib}
\docType{package}
\title{
Work with PTPd log file data.
}
\description{
A set of routines to read, graph and analyze PTPd log data.
}
\details{
\tabular{ll}{
Package: \tab ptplib\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-04\cr
License: \tab BSD_2_clause\cr
}
}
\author{
George Neville-Neil

Maintainer: George Neville-Neil <gnn@neville-neil.com>
}
\references{
http://ptpd.sf.net
}
\keyword{ package }
\seealso{
}
\examples{

}
