% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpOffsetStats}
\alias{ptpOffsetStats}
\title{Statistics on offset measurements}
\usage{
ptpOffsetStats(log, start, end, boundary = 1000)
}
\arguments{
\item{log}{dataframe returend by a ptpLogRead() call}

\item{start}{start time in strftime format}

\item{end}{end time in strftime format}

\item{boundary}{nanoseconds offset boundary}
}
\description{
Statistics on offset measurements
}

