% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpQualityGraph}
\alias{ptpQualityGraph}
\title{Graph quality data}
\usage{
ptpQualityGraph(logA, logB, output)
}
\arguments{
\item{logA}{data from host A}

\item{logB}{data from host B}

\item{output}{optional output file for the graph}
}
\value{
None
}
\description{
Graph quality data
}

