% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpQualityGraphCompare}
\alias{ptpQualityGraphCompare}
\title{Compare two quality graphs}
\usage{
ptpQualityGraphCompare(diffA, diffB, output)
}
\arguments{
\item{diffA}{difference A}

\item{diffB}{difference B}

\item{output}{optional output file for the graph}
}
\value{
None
}
\description{
Compare two quality graphs
}

