% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpQualityRead}
\alias{ptpQualityRead}
\title{Functions for working with PTPd quality files}
\usage{
ptpQualityRead(file)
}
\arguments{
\item{file}{quality file to read}
}
\value{
a data frame based on a PTPd quality file.
}
\description{
Functions for working with PTPd quality files
}

