% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpQualityStats}
\alias{ptpQualityStats}
\title{Statistics on quality measurements}
\usage{
ptpQualityStats(difference)
}
\arguments{
\item{difference}{dataframes returend by a ptpQualtiyRead() calls}
}
\description{
Statistics on quality measurements
}

