% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpStats}
\alias{ptpStats}
\title{Functions for deriving various statistics over a PTP log}
\usage{
ptpStats(log, start, end)
}
\arguments{
\item{log}{dataframe returned by a ptpLogRead() call}

\item{start}{start time in strftime format}

\item{end}{end time in strftime format}
}
\value{
None
}
\description{
Functions for deriving various statistics over a PTP log
}

