% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptplib.R
\name{ptpLogRead}
\alias{ptpLogRead}
\title{Read a PTP log file}
\usage{
ptpLogRead(file)
}
\arguments{
\item{file}{A properly formatted PTP log file generated by the daemon}
}
\value{
The raw data frame, merged timeseries (offset, delay, m->s, s->)
         offset time series, delay time series,
         master.to.slave time series, and slave.to.master time series
}
\description{
Read a PTP log file
}

